requirements.txt
flask
openai-whisper
gunicorn
librosa
soundfile
gpustat
grip
argparse
tqdm
service_streamer
requests
transformers
accelerate
datasets[audio]
pydub
